requirements.txt
backcall==0.1.0
bcrypt==3.1.6
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
cycler==0.10.0
decorator==4.3.0
Flask==1.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.5
ipython==7.2.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.2
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
kiwisolver==1.0.1
MarkupSafe==1.1.0
matplotlib==3.0.2
nbformat==4.4.0
newsapi-python==0.2.3
numpy==1.16.0
pandas==0.23.4
parso==0.3.2
pexpect==4.6.0
pickleshare==0.7.5
plotly==3.5.0
prompt-toolkit==2.0.7
protobuf==3.6.1
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.1
pyparsing==2.3.1
python-dateutil==2.7.5
pytz==2018.9
requests==2.17.1
retrying==1.3.3
scikit-learn==0.20.2
scipy==1.2.0
seaborn==0.9.0
six==1.12.0
sklearn==0.0
SQLAlchemy==1.2.18
traitlets==4.3.2
urllib3==1.21.1
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
